# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/edge/history": {
            "get": {
                "description": "取得所有歷史資料\u003cbr\u003e 若protocol=\"OPC UA\" Machine_Status(value=1:全自動,2:半自動,3:手動,4:設置,其他數字:Unknown) \u003cbr\u003e Machine_Motor(value=1:ON,0:OFF)\u003cbr\u003e 若protocol=\"HTTP\" Machine_Status(value=2:停機,3:閒置,4:異常,5:運轉,7:不良品,其他數字:Unknown)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歷史資料"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "選擇時間",
                        "name": "selectTime",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "顯示欄位",
                        "name": "columns",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/historyrouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/level0history": {
            "get": {
                "description": "取得level0歷史資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歷史資料"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "選擇時間",
                        "name": "selectTime",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/edgeservice.NodeStruct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/historyrouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/level1history": {
            "get": {
                "description": "取得level1歷史資料\u003cbr\u003e 若protocol=\"OPC UA\" Machine_Status(value=1:全自動,2:半自動,3:手動,4:設置,其他數字:Unknown) \u003cbr\u003e Machine_Motor(value=1:ON,0:OFF)\u003cbr\u003e 若protocol=\"HTTP\" Machine_Status(value=2:停機,3:閒置,4:異常,5:運轉,7:不良品,其他數字:Unknown)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歷史資料"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "選擇時間",
                        "name": "selectTime",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/edgeservice.NodeStruct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/historyrouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/level2history": {
            "get": {
                "description": "取得level2歷史資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歷史資料"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "選擇時間",
                        "name": "selectTime",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/edgeservice.NodeStruct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/historyrouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/level3history": {
            "get": {
                "description": "取得level3歷史資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歷史資料"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "選擇時間",
                        "name": "selectTime",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/edgeservice.NodeStruct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/historyrouter.errMsg"
                        }
                    }
                }
            }
        },
...
        "/edge/level4history": {
            "get": {
                "description": "取得level4歷史資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "歷史資料"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "設備ID",
                        "name": "id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "選擇時間",
                        "name": "selectTime",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/edgeservice.NodeStruct"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/historyrouter.errMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "edgeservice.NodeStruct": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "object"
                }
            }
        },
        "historyrouter.errMsg": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}